Tweaks to allow proper consumption from NPM & Typescript
