+++ tsconfig.json
@@ -11,7 +11,7 @@
     "experimentalDecorators": true,
     "forceConsistentCasingInFileNames": true,
     "importHelpers": true,
-    "module": "AMD",
+    "module": "CommonJS",
     "moduleResolution": "node",
     "noFallthroughCasesInSwitch": true,
     "noImplicitOverride": true,
@@ -20,7 +20,6 @@
     "noUncheckedIndexedAccess": true,
     "noUnusedParameters": true,
     "outDir": "./dist",
-    "outFile": "./dist/index.js",
     "sourceMap": true,
     "strict": true,
     "target": "ES2022",
